Implement DOM manipulation content script (#3)
* Initial wxt implementation

* Improve DOM update logic, add `hidePost` function

Previously there were dozens of redundant processing cycles. This commit
processes specific DOM element updates and parses each one individually
rather than parsing an entire collection at once. Now, only each new
tweet element that is added into the DOM is processed accordingly.

Testing is being performed with an unauthenticated incognito session on
a user's profile page. The observed behavior removes and adds tweet
elements into the DOM during page scroll. With the new logic, each
tweet is processed as Twitter's logic adds it into the DOM.

We also add a first implementation of `hidePost`, which adds a custom
class to tweets that exceed the negative threshold. It seems the
`display: none;` is not being applied (blocked by browser?); will need
to figure this out.

* `Parser` updates, add `constants.ts`

Added the `RANK_API` constant for now.

* Twitter: correctly parse `postId`

Previous behavior would always pull the postId as the last URI of the
`/status` path, but sometimes there are additional URIs appended to the
full URL path. This commit grabs the `postId` as the URI that follows
the `/status` URI.

* Improve DOM manipulation

This commit introduces a significant performance and logic improvement
to the DOM manipulation code.

The `MutationObserver` now watches the `react-root` node, rather than
awaiting a specific `timeline` node, which avoids the need to watch for
URL changes as the `react-root` will not change even with SPA/HTML5 path
changes. This alone increases runtime performance dramatically and avoids
race conditions

Timeline ads are also now detected and filtered automatically :)

* Cleanup, name changes

Change the `MutationObserver` from `root` to `react-root` to match the
Twitter node name and also change console logging to say `post` rather
than `tweet`.

* Update Selectors

Several jquery selectors have been added/fixed in order to handle more
cases, e.g. notifications vs tweets, retweet author vs original tweet
author, etc.

* Fix Parser for `postId`, `userName` -> `profileId`

Parsing for `postId` would grab the last chunk of the URL as the `postId`
but this does not always work. Now the parser will find the `/status/`
chunk and take the next chunk as the `postId`.

* Improve structure/logic of Twitter content-script

This commit makes several significant changes to the structure and logic
of the Twitter DOM manipulation code.

- Individual functions were moved into a `DOM` class
- `isBelowThreshold` method now returns a boolean; calling method hides post
-  `retweetProfileId`s are now properly parsed; processing is todo
- Context was updated to run at `document_idle` to ensure DOM `body` element is loaded before attaching `MutationObserver`

* Fix false positives in ad detection

Previous logic was catching twitter names and tweet text containing the
string "ad". This commit narrows the selection to the `span` element
that contains the "Ad" string.

It's possible there are some other unforeseen cases that will trigger
false positives, but for now this selector is working correctly.

* Prettier, log `elementType` for processing

* Prettier formatting