Cleanup and updates (#9)
* Add platform namespaces for DOM utils

Organization for future growth

* `stores/wallet.ts`: assert item key before value

It makes more sense to ensure the key->value store has a key first before
asserting the value.

* Comments, logging

* Initialize `local:wallet:utxos` as empty Map

* `hasSeedPhrase`: return result as boolean

Previously this method was a ternerary operator; now it instantiates the
result into a `Boolean` and returns that.

* `modules/wallet.ts`: Prettier

* Adjust `WalletManager` init sequence

This commit moves the Chronik initializers to the `WalletManager.init()`
method and converts `WalletManager.init()` to an async method.

* Adjust `WalletManager.outpoints` parse behavior

Previous behavior would convert the `utxos` Map to an array to `map()`
through the new array and return the `OutPoint[]` data.

This commit adjusts the behavior to loop through the Map as-is to compile
and return the `OutPoint[]` data.

* `WalletManager`: remove `onWsConnect` method

No need to have this method just to log to console. Console logging was
added as an anonymous arrow function to the `ws.onConnect` event handler.

* Log exception errors for `popup:seedPhrase` handler

* Use new platform selector/parser namespaces

Should have been included in 826c451

* Remove unneeded `wss` host permission from manifest

* Call `startMutationObserver` from `new DOM()`

No need to instantiate this class to a constant; just call the runtime
mutation observer directly from a new instantiation.

* Add UI dependencies, update/`--save-dev` existing

Update all existing dependencies and move all (except for `vue`) into
devDependencies. Also install `primevue` to prepare for the UI buildout.