bugfix: load content-script at `document_end`
It is necessary to run the content-script after the DOM has been fully
loaded, otherwise the constructor function for the `Mutator` class will
throw and exception and the content-script will not initialize. This is
because the `react-root` div needs to be loaded first so that the
`MutationObserver` can be attached to it for monitoring Twitter DOM
mutations during runtime.