Twitter integration (#15)
* Full Twtter implementation

This commit fully implements the Twitter functionality. It is not a
complete implementation, but a prototype that provides the baseline
functionality.

- Upvote/downvote buttons on all posts
- Upvote/downvote buttons will cast votes by burning Lotus
- Update vote count and post rankings in realtime
- Hide posts that are below the default ranking threshold of 0

Will still need to provide a mechanism that will periodically fetch
post rankings so that other people's votes will be accounted for in
realtime.

* Fetch total post ranking/votes after casting vote

This commit changes the behavior of the vote handling function so that
the post that was voted on will be updated to reflect the total vote
count. This is accomplished by fetching the post ranking and vote count
from the RANK API after a vote has been cast.

* Add/remove functions

Add a new `blurPost` function that will be used to blur out posts that
have a negative reputation when they appear in the timeline.

Remove the `isPostBelowThreshold` function as this was using the vote
count to determine ranking, but this is not ideal long-term when users
have the option to change their default vote amount (currently 100 Lotus)

* bugfix: load content-script at `document_end`

It is necessary to run the content-script after the DOM has been fully
loaded, otherwise the constructor function for the `Mutator` class will
throw and exception and the content-script will not initialize. This is
because the `react-root` div needs to be loaded first so that the
`MutationObserver` can be attached to it for monitoring Twitter DOM
mutations during runtime.

* Tons more optimization

- Added new Map for caching profile ranking data for future avatar badges
- `RegExp` to detect post URLs to not blur downranked thread context
- Adjusted some upvote/downvote button attributes for cleaner processing
- Added the blur overlay button + styling
- Added blur click action (click the article element to navigate to
post status page as though the actual post itself were clicked)
- Reverted post ranking and vote count update after button click; this
functionality will be moved to the API polling interval in the future
- Only hide Ads; blur downranked posts
- Cache the post article element rather than the `cellInnerDiv` wrapper div
- CSS stylesheet for `blurred-overlay`

- Miscellaneous logic improvements, comments, etc.

* CSS tweaks

* Update selectors

Add profile avatar and Grok actions button selectors

* Update CSS for post blurring

The `.blurred` class now cascades through all decsendants and does a much
better job at blurring out posts. Opacity and box-shadow properties were
also updated to coincide with the new cascading.

* Tons of changes to Twitter `content-script`

Lumping all changes into this one commit; too many to sift through at
the moment.

* Use `checkPostRankingAndBlur` everywhere

Added the function but neglected to use it in all necessary places.

* Significant logic improvements

This commit changes function names and adjusts the logic centered around
cached post processing/updating.

* Disable vote button when clicked, re-enable after

When a vote button is clicked on a post, first disable the button to
prevent additional processing. Once the processing is completed, the
vote button is re-enabled.

* `validateMutatedElement` -> `isValidElement`

Also add a comment for `attributeFilter`, which we may use in the future
to streamline post/profile processing.