Adjust `.blurred` and `.blurred-overlay` classes
`.blurred` was adjusted to pass the blur filter to all descendant elements,
while the remaining CSS properties were passed specifically to the `div`
element that contains the `article` element.

The `.blurred-overlay` class was also adjusted to widen the button so
that the child `span` element doesn't wrap on smaller screens.