Adjust `.blurred` class for performance on mobile (#20)
Previous tests proved to be a disaster on mobile. It seems the cause was
the blurring effect cascading through all of the subelements of the
article and was overloading the mobile rendering.

This commit reduces the cascading to the `article`'s parent div, and then
applies only to the parent `div` elements of `img` elements or video
rendering `div` elements.

Tests on mobile show a significant improvement in performance. Other
adjustments may need to be made in the future, but this is a good start.