Wallet messaging, performance, stability tweaks (#22)
* Add WebSocket ping interval

This is to help prevent the background service-worker from going inactive
on Chromium-based browsers.

* Remove `EventCallback`, return `txid` directly

This commit adjusts the background service-worker logic for RANK and
send transaction crafting and broadcasting. The logic has been
drastically simplified such

1. the service-worker will receive the request from the extension to
craft a transaction
2. the service-worker will call the handler in the `WalletManager`
3. the handler will craft the transaction and broadcast it
4. if successful, the `txid` will be returned to the service-worker, and
then returned to the popup UI or content-script

Along with this adjustment comes disabling the `reconcileUtxos` method.
Instead, the transaction crafting logic now returns the `OutPoint`s
which are used in the transaction. If the transaction is broadcasted
successfully, the spent `OutPoint`s are then processed through the
`WalletManager.outpoints` setter function, which deletes the UTXO from
the `UtxoCache` and updates the runtime balance.

* Add comment blocks

* Adjust `WalletMessaging.popup:sendLotus` return value

With the recent change, the return value of this `WalletMessaging`
interface needs to be changed for the appropriate return value. We now
await the `Promise` for the `txid` string value.

* Designate `WalletManager` methods as private

No need to have these public anymore at this point.