Significant improvements to Twitter content-script
This commit vastly improves the UX on Twitter by fixing the vote button
SVGs to fit better, and also adds profile avatar badges on posts to indicate
a profile's overall reputation (red = bad, green = good, grey = neutral).
This necessitates the use of the `ProfileCache` map and ensures that the
cached profiles are updated when a profile's post rankings are updated from
the API. This commit also sets up some skeleton code to better handle profile avatar
badging for a future version. For now, only posts will have the profile
avatars badged with the profile's reputation.

The logic for the `postVoteUpdateInterval` was drastically improved to
removed posts from the `PostCache` if they are no longer present in the
DOM, and also to balance the timing of the interval based on how many
posts are cached.