Twitter and Wallet improvements (#23)
* twitter: fix vote arrow SVGs

* WalletMessaging: adjust return value definitions

Adjust return values for transaction crafting message handlers to be
`void` rather than `null`. Also throw errors rather than logging to
console, as they will get logged to the console by the calling function.

* WalletManager: performance and logic improvements

Changes in this commit:
- Remove `WalletBuilder` interface definition
- Add logic to better recover UTXO state from Chronik WebSocket errors
- Add processing logic to `wsPingInterval` and add UTXO hydration logic
- Adjust logic to reduce the propensity for corruption of the
`MutableWalletState`
- Prettier formatting

* `background/index.ts`: comments, error-handling

Add comment blocks and wrap the `popup:loadSeedPhrase` handler code in
a `try/catch` block.

* Popup UI title -> "Lotus Voting System"

* Twitter: Add skeleton storage API classes

This will be helpful for caching profile and post data to improve UX

* Twitter: add more selectors and parsers

* Twitter: adjust CSS styling

Adjust the `blurred-overlay` and the profile avatar badges

* `utils/functions.ts`: small tweaks

* Significant improvements to Twitter content-script

This commit vastly improves the UX on Twitter by fixing the vote button
SVGs to fit better, and also adds profile avatar badges on posts to indicate
a profile's overall reputation (red = bad, green = good, grey = neutral).
This necessitates the use of the `ProfileCache` map and ensures that the
cached profiles are updated when a profile's post rankings are updated from
the API. This commit also sets up some skeleton code to better handle profile avatar
badging for a future version. For now, only posts will have the profile
avatars badged with the profile's reputation.

The logic for the `postVoteUpdateInterval` was drastically improved to
removed posts from the `PostCache` if they are no longer present in the
DOM, and also to balance the timing of the interval based on how many
posts are cached.