Complete overhaul of Twitter DOM processing
Primarily, the overhaul comes from implementing jQuery to handle all
DOM processing operations. With this, much of the other processing and
storage logic was conflicting and/or needed to be changed/removed.

Profile avatar badges are set on any page that includes an avatar
and also has the requisite profile/post data needed to fetch the profile
reputation from the API.

Also, vote buttons are not only added into post elements. They are
added into any button row element that includes the original "Like"
action button. On mobile, when viewing a photo or video, you can cast
votes while viewing, rather than having to return to the post element
to do so.