Overhaul Twitter implementation + `WalletManager` tweaks (#26)
* Add and adjust selectors and CSS classes

Added some additional CSS classes to handle more profile avatar element
pixel sizes. Added some additional selectors for finding profile avatars
on more pages and in more containers within the observed DOM element.

* Remove `cheerio`, install `jquery`

jQuery is preferred in this instance since it is able to work directly
with DOM elements, whereas cheerio is not. This will be useful for
streamlining all DOM processing logic.

* More selector updates

Tons of changes made after testing with the new jQuery implementation.
jQuery commits to follow.

* Update parser to use 'jquery', remove 'cheerio'

* Fix vote arrow SVG dimensions/layout

Previously the SVG elements were slightly off-center. This commit fixes
the `path` for the SVGs to work better within the post button rows.

* Complete overhaul of Twitter DOM processing

Primarily, the overhaul comes from implementing jQuery to handle all
DOM processing operations. With this, much of the other processing and
storage logic was conflicting and/or needed to be changed/removed.

Profile avatar badges are set on any page that includes an avatar
and also has the requisite profile/post data needed to fetch the profile
reputation from the API.

Also, vote buttons are not only added into post elements. They are
added into any button row element that includes the original "Like"
action button. On mobile, when viewing a photo or video, you can cast
votes while viewing, rather than having to return to the post element
to do so.

* Update Twitter constants

* Wallet balance/utxo stability improvements

With mobile testing, issues with the UTXO state would frequently arise
due to Chronik websocket disconnects. This commit aims to provide better
resilience and runtime stability to avoid corruption of the UTXO state.

* `wallet.ts`: prettier formatting