More Twitter CSS adjustments
This commit uses percentages for badge sizing and placement. All badges
are now rendering very nicely. Small avatars (i.e. <24px w/h) are not
rendered with badges as the avatars are generally rendered without profile
data attached to them, and are usually rendered in a row with other avatars,
so adding badges is not very helpful in these scenarios.