Twitter: add post cache expiry time
This commit adds an expiry to cached posts, which helps to reduce the
number of requests made to the API, and also stabilizes the UX as the
client does not need to always request ranking data for a post as the
user is navigating between posts/profiles.

The default expiry time is 5s, which is also used in the automated post
update interval. This time may be extended in the future, but 5s works
well in my testing.