Inching closer to beta (#28)
* Add Lotus icons for extension package

Replace the default WXT icon

* More adjustments to Twitter vote button SVGs

The iteration in this commit is very good now. I am satisfied. :)

* More Twitter CSS adjustments

This commit uses percentages for badge sizing and placement. All badges
are now rendering very nicely. Small avatars (i.e. <24px w/h) are not
rendered with badges as the avatars are generally rendered without profile
data attached to them, and are usually rendered in a row with other avatars,
so adding badges is not very helpful in these scenarios.

* Twitter: add post cache expiry time

This commit adds an expiry to cached posts, which helps to reduce the
number of requests made to the API, and also stabilizes the UX as the
client does not need to always request ranking data for a post as the
user is navigating between posts/profiles.

The default expiry time is 5s, which is also used in the automated post
update interval. This time may be extended in the future, but 5s works
well in my testing.

* Twitter: add/update jQuery selectors

More testing and refinement to existing selectors, and also added new
ones for additional functionality that will be included in a later
release.

* `@utils/functions`: Add `toMinifiedNumber()`

This function will be useful to minify large numbers, e.g. to display
profile rankings on a profile's page

* Massive improvements to Twitter content-script

This is a big one. Highlights for this commit include:

- Enforce cached post expiry time
- Remove the `Mutator` class and replace it with two new classes:
`AddedMutator` and `RemovedMutator`. These new classes are each called
during an appropriate mutation event, either when nodes are added or
removed from the DOM.
- Overhaul mutation processing logic: Previous logic was checking for
a mutation to match one of our desired processing conditions, however
this was beginning to become a scalability issue due to mutations
carrying more than one mutation we wanted to process. This commit overhauls
the entire logical flow, where each mutation is checked against every
one of our processing conditionals, and only those that match are
processed. Due to this change, duplicate processing has also been reduced
to near 0% (some avatars still get processed more than once in some
scenarios)
- Several processing and rendering bugs fixed with new mutation processing
logic
- Removed unneeded types and added some new ones
- Preparations for displaying vote counts and overall ranking on profile
pages
- Probably a couple miscellaneous things I forgot to mention

This update was all about performance and reliability.

* Update dependencies

* Apply prettier formatting

Some files needed formatting changes applied via `npx prettier -w .`