bugfix: ignore button rows containing vote buttons
In rare cases on mobile browser (Android, Kiwi browser) that button row
divs on some posts would be processed twice, leading to an additional
downvote button appearing on the post. It is likely a race condition
between the page loading delays and the mutation observer.

This commit adds a conditional to the `processButtonRows` mutation handler
that will check for the existence of the Lotus vote buttons and skip
elements that contain them.