Fix race condition, ignore unneeded DOM mutations (#30)
* bugfix: ignore button rows containing vote buttons

In rare cases on mobile browser (Android, Kiwi browser) that button row
divs on some posts would be processed twice, leading to an additional
downvote button appearing on the post. It is likely a race condition
between the page loading delays and the mutation observer.

This commit adds a conditional to the `processButtonRows` mutation handler
that will check for the existence of the Lotus vote buttons and skip
elements that contain them.

* Skip processing for some element types

This commit adds a conditional check to `handleMutatedNode` that will
skip several unnecessary mutation events that we don't need to spend
CPU cycles processing. The main processing function `processMutatedElement`
is then called afterwards for any elements of interest.