bugfix: button rows sometimes lack post URL element
It was observed that some button rows would be rendered without the
vote buttons. This would occasionally happen when viewing photos on a
smaller screen; the media viewer would render the photo without a child
element that contains the `<a>` element that was expected for gathering
the post URL, hence the `profileId` and `postId` would not be found and
the button row mutation would fail.

This commit changes the logic to use the `window.location.pathname` as
the source of the post URL if there is no `<a>` child element to parse
for this.