Bugfixes and tweaks (#32)
* fix avatar badge class detection for Safari

It was observed that Safari could not parse the existing `reputation`
class name on avatar elements with the `entries().find()` method that was
implemented. This commit replaces that method with one that works in
both Chrome and Safari.

* bugfix: button rows sometimes lack post URL element

It was observed that some button rows would be rendered without the
vote buttons. This would occasionally happen when viewing photos on a
smaller screen; the media viewer would render the photo without a child
element that contains the `<a>` element that was expected for gathering
the post URL, hence the `profileId` and `postId` would not be found and
the button row mutation would fail.

This commit changes the logic to use the `window.location.pathname` as
the source of the post URL if there is no `<a>` child element to parse
for this.

* Adjust wallet initialization logic

This commit changes the `WalletManager` initialization logic to load
the local storage state during `WalletManager.init()`. The constants
are also moved inside the defined background script `main()` function
to ensure a consistent JavaScript scope across the background service
worker.

* `WalletManager`: Adjust `wsPingInterval` logic

Adjustments to further help avoid the background service worker from
going inactive. This should help with Safari browsers.

* Twitter: simplify avatar badges

Adjust the `forEach()` loop a bit. Nothing crazy.

* Twitter: `updateCachedPosts` vote button updates

Process each vote button in a loop rather than procedurally.

* Twitter: process button row elements asynchronously

Rather than jQuery `.each()`, use `.map()` for asynchronous callback
function.

* `WalletManager`: check wallet state on tx craft

It was observed that the wallet would not submit a transaction due to
missing txins. This commit adds a check to ensure the `walletState` is
loaded, in case the background service worker was suspended prior to
receiving the message from the content script.

* Update `buttonRow` jQuery selector

To coincide with the button row processing changes from previous commits.