More bugfixes and tweaks (#35)
* bugfix: properly block ads

Previous ad selector wasn't blocking the first ad displayed in the timeline.
This commit fixes ad-blocking logic.

* Adjust Grok selectors and handlers

This commit adds the "Grok Drawer" element and renames the appropriate
handler method.

* Adjust button `.on()` listener instantiation

Previous logic would pass the raw `HTMLElement` to the method. This commit
changes the instantiation to remove the need for passing the raw element.

* bugfix: auto-update post vote counts failing

A regression was introduced in v0.3.2 that broke the auto-update feature
for posts, which would not properly update the post vote counts and would
also not use the updated `CachedPost` data when processing vote counts.

This commit fixes both of those broken behaviors.

* Process primary/sidebar columns

This commit adjusts the column processing logic to add selectors to an
array, then joins that array with the "`, `" string to find the mutated
elements during processing.

* Skip more mutations

* Add comments and JSDoc

* bugfix: optional chaining for `outpoints` getter

This commit fixes unhandled exceptions (encountered when checking `this.outpoints`
during transaction processing) by doing optional chaining.

* `WalletManager.deinit()`: clear interval first

* `WalletManager`: adjust WebSocket health check

Check to see if the WebSocket status is closed/closing rather than open

ref: https://stackoverflow.com/questions/23369368/how-to-get-the-current-status-of-a-javascript-websocket-connection

* `WalletManager`: await `processEventQueue`

Await resolution of the promise, rather than returning.

* `wxt.config.ts`: add Safari runner

* Update dependencies