Twitter: significant CSS optimizations
All of the profile avatar sentiment classes were remvoed in favor of
custom variables to apply conditionally. Also, the `.blurred` class adds
all of the styling for the overlay and the button itself.

This commit also adds CSS button styling to indicate the extension instance
has voted on a post.