Implement voting history, tracking (#38)
* Update `rank-lib` submodule

* Add `instance` store

This is used for tracking instance state and vote history.

* Export instantiated `instanceStore`, types

* Twitter: significant CSS optimizations

All of the profile avatar sentiment classes were remvoed in favor of
custom variables to apply conditionally. Also, the `.blurred` class adds
all of the styling for the overlay and the button itself.

This commit also adds CSS button styling to indicate the extension instance
has voted on a post.

* Twitter: deleted unneeded store modules

The post metadata cache was implemented as a background store.

* background: add `getScriptPayload` handler

The Twitter content script will need this data to submit personalized
API requests that return post metadata associated with the wallet vote.

* background: save instance data on init

* background: cleanup wallet initialization

* background: loadWalletState: return state directly

Rather than executing another messaging call, return the `uiWalletState`
directly, which is then used in App.vue.

* `WalletMessaging`: add `scriptPayload` getter

Also adjust the return values for the other message handlers that were
adjusted.

* `utils/functions`: add `newInstanceId`

This is a small proof-of-work method that generates an `id` for this
instance of the extension. This will be useful later for registering
with the Lotusia Stewardship for participating in rewards programs.

* `WalletStore`: adjust schema, add item definitions

Necessary changes to ensure the required data for the `instanceStore` and
content-scripts is available.

* `WalletManager`: add `scriptHex`, `scriptPayload`

* Add, update Twitter jQuery selectors

* `.prettierrc`: return to default width of 80

* Apply prettier formatting

* Twitter: support scriptPayload requests, `PostMetaCache`

This commit adds all of the support for sending API requests accompanied
by the wallet's `scriptPayload`. The content-script then caches the
`PostMeta` content into localStorage.

There are also some miscellaneous improvements and bugfixes.

* `WalletManager`: remove outpoints check

This commit removes the deprecated check that was called prior to crafting
and submitting RANK/send transactions.

* `WalletStore`: remove `assert` call for item value

* `WalletStore`: add `addressStorageItem` getter

* background: check wallet endpoints before `RANK` tx

* background: prettier formatting for imports

* `App.vue`: add `watchers` key for `address`

Will be helpful once seed import is a feature.

* Twitter: join selector array with `', '`

* `WalletStore`: add some jsdoc

* `HomeNav.vue`: prettier formatting

* More prettier formatting

* wxt.config.ts: disable `transformMixedEsModules`

* `WalletManager`: init `script` from `scriptHex`

* `WalletManager`: conditional for empty UTXO set