Add seed import feature (#42)
* move `HomeNav.vue` content to `Footer.vue`

Currently there is no need to have the `HomeNav` content nested within
the footer, so it was moved to the `Footer` component. This is helpful
so that we don't need to layer the emitted events.

* `WalletBuilder`: add `isValidMnemonic` method

This method validates seed phrases when a user attempts to import it into
the wallet.

* `WalletManager`: add optional chaining for getters

With the new import logic, `this.wallet` may not be defined, so getters
need optional chaining to prevent unhandled exceptions.

* `instanceStore`: add `getInstanceId` method

This is used to ensure we don't generate a new `instanceId` when a new
seed phrase is imported.

* Don't generate `instanceId` in background

This is being moved to `App.vue`

* Adjust QR code size

240px down to 200px

* Adjust popup CSS

* `App.vue`: remove PrimeVue import/config

* Implement seed phrase import

Add the necessary Vue components and make necessary adjustments to
background service worker to implement the new feature.

* Move `instanceId` generation to background

Moving this back to the background service worker. The `instanceStore`
is referenced and a new `instanceId` is generated if one does not already
exist.