Some fixes and updates (#44)
* Simplify wallet setup, disconnect handling

Some tweaks to the Chronik (re)setup that should help with stability

* Overwrite `postMeta` when available and specified

If we are specifying the `updateCachedPost` function to save `postMeta`,
then we do it without caring if it already exists.

* Verify `voteButtons` exist before updating vote count

This fixes a bug where DOM `button` elements that were not RANK vote buttons
would have a `span` element appended to them with a vote count.

* Tweak avatar badge CSS, use pixel width number ranges

Adjust the avatar badge CSS to accommodate some evident changes to
Twitter's DOM elements that appeared over this past week. Also change
the conditional logic in `setProfileAvatarBadge` to apply CSS classes
if the avatar width is within a range of pixels, rather than a static
pixel width.

* Adjust Parser, simplify `parsePostElement` logic

Small adjustments to the `Parser`, and vast simplification of the
`parsePostElement` logic.

* Improve jQuery function call chaining

Don't use the `:first`/`:last` selectors; instead use `.first()`/`.last()`
methods of the jQuery objects since the selectors are deprecated since
jQuery 3.4.

* Better filtering button row for "(un)like" button

The like/unlike buttons have a `data-testid` attribute while most other
buttons do not. This commit adds the `data-testid` condition to reduce
improve filtering accuracy.

* Minor cleanup

Comments, conts, and imports. Oh my!

* Function updates

* Update `rank-lib` to 9f148b5