Implement instance registration with RANK API
This branch adds all of the requisite modules and runtime logic to
register new extension installations with the RANK API backend. This
registration is tracked on the backend to determine which installations
will receive news bulletins, eligibility for engagement rewards, etc.

The implementation in the popup UI is not pretty, but it is functional and
will be further refined in later commits.