Revamp popup UI
This branch brings significant UI improvements to the popup UI, and adds the critical, missing functionality to send Lotus to other wallets. With the UI changes, the structure of the Vue components has been modified and improved significantly.

This branch also implements a basic "trending" feed that fetches the top 5 profiles/posts that are currently trending for the day.