bugfix: extension popup UI dimensions (#51)
This commit fixes a bug introduced in v1.0.0-beta where the `window.top`
and `window.left` properties can both be `0` depending on the monitor
on which the browser application is rendered.

This commit adds a check for `window.state` to ensure that the hard-coded
popup UI dimensions are set properly.