Fixup HTML layout, optimize runtime (#71)
This commit fixes a long-standing issue with the HTML layout. Now the header (balance), footer (navigation), and individual page content are rendered and stickied properly. Also the pages are now scrollable! Now we can add a bunch more stuff for people to do, and they'll be able to scroll to it. :)

This commit also adds the "My Stats" page and properly implements authorized fetching from the Lotusia RANK API.

Along with these changes were some minor adjustments to the Vue component initialization strategy for the Main layout. Now the pages are only rendered once the required reactive values have been hydrated via runtime operations or storage.