Initialize `chainState` on wallet init
When the wallet is first built and initialized, the `chainState` is set to default values. This is problematic because the popup now relies on hydrated `chainState` values for runtime.

This commit adds a method to the `WalletManager` that is called upon `init()` to ensure the initial `chainState` values are hydrated on first-launch.