More adjustments, optimizations (#77)
This commit removes `BlockDataSig` storage from the instance store. It also ensures that a fresh authorization header is generated whenever a new wallet password is imported. This is to coincide with changes to the backend API that validates the `scriptPayload` that signed the authorization payload matches the request being made.

The logic flow for authorized API calls was also improved, and a new component was added to show the loading spinner with the provided message when awaiting data.