Minor fixups, prepare UTXO consolidation
This commit sets up the required Wallet module/messaging for consolidating the UTXO set if it grows too large such that all of the Lotus in the wallet cannot be sent in a single transaction. There is also some minor cleanup in the Main component, removing the unneeded storage watcher for the current blockchain tip/height.