Use UTXO consolidation, other tweaks
This commit adds the necessary logic to ensure that the wallet UTXO set is consolidated after the `WalletManager` is initialized. UTXO consolidation is performed if the wallet's entire UTXO set cannot be spent in a single transaction.

This commit also makes some small adjustments and cleanup. Also fixes the popup window size on mobile Chrome browsers.