Implement customizable social settings, adjust Twitter content-script and CSS (#83)
This commit adds a brand new `SettingsStore` class for customizing the RANK configuration of the extension. These new settings are also implemented in the Twitter content-script.

There are also some other minor changes added to the injected Twitter CSS to enable a gradient of colors for the profile avatar reputation badges.